+++ R/imputation.R
@@ -55,8 +55,8 @@
 ##' the most similar features to impute a feature with missing values,
 ##' or the most similar samples to impute all missing in a sample.
 ##'
-##' The `margin` argument can be used to change the imputation margin
-##' from features/rows (`margin = 1`) to samples/columns (`margin = 2`).
+##' The `MARGIN` argument can be used to change the imputation margin
+##' from features/rows (`MARGIN = 1`) to samples/columns (`MARGIN = 2`).
 ##' Different imputations will have different default values, and
 ##' changing this parameter can have a major impact on imputation results
 ##' and downstream results.
@@ -215,7 +215,7 @@
 ##' @param ... Additional parameters passed to the inner imputation
 ##'     function.
 ##'
-##' @param margin `integer(1)` defining the margin along which to
+##' @param MARGIN `integer(1)` defining the margin along which to
 ##'     apply imputation, with `1L` for rows and `2L` for columns. The
 ##'     default value will depend on the imputation method. Use
 ##'     `getImputeMargin(fun)` to get the default margin of imputation
@@ -305,21 +305,21 @@ imputeMethods <- function()
 ##' @param k `numeric(1)` providing the imputation value used for the
 ##'     first and last samples if they contain an `NA`. The default is
 ##'     to use the smallest value in the data.
-impute_neighbour_average <- function(x, k = min(x, na.rm = TRUE), margin = 1L) {
+impute_neighbour_average <- function(x, k = min(x, na.rm = TRUE), MARGIN = 1L) {
     message("Assuming values are ordered.")
-    margin <- .checkMargin(margin)
-    if (margin == 2L) x <- t(x)
+    MARGIN <- .checkMargin(MARGIN)
+    if (MARGIN == 2L) x <- t(x)
     res <- .Call("C_impNeighbourAvg", x, k)
-    if (margin == 2L) return(t(res))
+    if (MARGIN == 2L) return(t(res))
     else return(res)
 }
 
 ##' @export
 ##' @rdname imputation
-impute_knn <- function(x, margin = 1L, ...) {
+impute_knn <- function(x, MARGIN = 1L, ...) {
     requireNamespace("impute")
-    margin <- .checkMargin(margin)
-    if (margin == 2L)
+    MARGIN <- .checkMargin(MARGIN)
+    if (MARGIN == 2L)
         x <- t(x)
     imp_res <- impute::impute.knn(x, ...)
     if (!is.null(imp_res$rng.state)) {
@@ -327,38 +327,38 @@ impute_knn <- function(x, margin = 1L, ...) {
     } else {
         rm(".Random.seed", envir = .GlobalEnv)
     }
-    if (margin == 2L)
+    if (MARGIN == 2L)
         return(t(imp_res$data))
     else imp_res$data
 }
 
 ##' @export
 ##' @rdname imputation
-impute_mle <- function(x, margin = 2L, ...) {
+impute_mle <- function(x, MARGIN = 2L, ...) {
     requireNamespace("norm")
-    margin <- .checkMargin(margin)
-    if (margin == 2L)
+    MARGIN <- .checkMargin(MARGIN)
+    if (MARGIN == 2L)
         x <- t(x)
     s <- norm::prelim.norm(x)  ## preliminary manipulations
     th <- norm::em.norm(s, ...) ## find the MLE
     seed <- sample(.Machine$integer.max, 1)
     norm::rngseed(seed) ## set random number generator seed
     res <- norm::imp.norm(s, th, x)  ## impute missing data under the MLE
-    if (margin == 2L)
+    if (MARGIN == 2L)
         res <- t(res)
     res
 }
 
 ##' @export
 ##' @rdname imputation
-impute_mle2 <- function(x, margin = 2L, ...) {
+impute_mle2 <- function(x, MARGIN = 2L, ...) {
     requireNamespace("norm2")
-    margin <- .checkMargin(margin)
+    MARGIN <- .checkMargin(MARGIN)
     dn <- dimnames(x)
-    if (margin == 2L)
+    if (MARGIN == 2L)
         x <- t(x)
     res <- norm2::emNorm(obj = x, ...)$y.mean.imp
-    if (margin == 2L)
+    if (MARGIN == 2L)
         res <- t(res)
     dimnames(res) <- dn
     res
@@ -366,10 +366,10 @@ impute_mle2 <- function(x, margin = 2L, ...) {
 
 ##' @export
 ##' @rdname imputation
-impute_bpca <- function(x, margin = 1L, ...) {
+impute_bpca <- function(x, MARGIN = 1L, ...) {
     requireNamespace("pcaMethods")
-    margin <- .checkMargin(margin)
-    if (margin == 2L)
+    MARGIN <- .checkMargin(MARGIN)
+    if (MARGIN == 2L)
         x <- t(x)
     nSamples <- dim(x)[2]
     .resultBPCA <- pcaMethods::pca(x,
@@ -378,20 +378,20 @@ impute_bpca <- function(x, margin = 1L, ...) {
                                    verbose = FALSE,
                                    ...)
     res <- pcaMethods::completeObs(.resultBPCA)
-    if (margin == 2L)
+    if (MARGIN == 2L)
         res <- t(res)
     res
 }
 
 ##' @export
 ##' @rdname imputation
-impute_RF <- function(x, margin = 2L, ...) {
+impute_RF <- function(x, MARGIN = 2L, ...) {
     requireNamespace("missForest")
-    margin <- .checkMargin(margin)
-    if (margin == 1L)
+    MARGIN <- .checkMargin(MARGIN)
+    if (MARGIN == 1L)
         x <- t(x)
     res <- missForest::missForest(x, ...)$ximp
-    if (margin == 1L)
+    if (MARGIN == 1L)
         res <- t(res)
     res
 }
@@ -410,7 +410,7 @@ impute_RF <- function(x, margin = 2L, ...) {
 ##' @export
 ##'
 ##' @rdname imputation
-impute_mixed <- function(x, randna, mar, mnar, margin = 1L, ...) {
+impute_mixed <- function(x, randna, mar, mnar, MARGIN = 1L, ...) {
     if (missing(randna))
         stop("Mixed imputation requires 'randna' argument. See ?impute_mixed.",
              call. = FALSE)
@@ -421,16 +421,16 @@ impute_mixed <- function(x, randna, mar, mnar, margin = 1L, ...) {
     if (missing(mnar))
         stop("Mixed imputation requires 'mnar' argument. See ?impute_mixed.",
              call. = FALSE)
-    suppressMessages(margin <- .checkMargin(margin))
-    if (margin == 2L) x <- t(x)
+    suppressMessages(MARGIN <- .checkMargin(MARGIN))
+    if (MARGIN == 2L) x <- t(x)
     if (length(randna) != nrow(x))
         stop("Number of proteins and length of randna must be equal.",
              call. = FALSE)
     x[randna, ] <- impute_matrix(x[randna, ], mar,
-                                 margin = margin, ...)
+                                 MARGIN = MARGIN, ...)
     x[!randna, ] <- impute_matrix(x[!randna, ], mnar,
-                                  margin = margin, ...)
-    if (margin == 2L) x <- t(x)
+                                  MARGIN = MARGIN, ...)
+    if (MARGIN == 2L) x <- t(x)
     x
 }
 
@@ -452,10 +452,10 @@ impute_min <- function(x) {
 ##' @param q `numeric(1)` indicating the quantile to be used to
 ##'     estimate the minimum in `MinDet` and `MinProb`. Default is
 ##'     0.01.
-impute_MinDet <- function(x, q = 0.01, margin = 2L) {
-    margin <- .checkMargin(margin)
-    n <- dim(x)[margin]
-    impVals <- apply(x, margin, quantile, prob = q, na.rm = TRUE)
+impute_MinDet <- function(x, q = 0.01, MARGIN = 2L) {
+    MARGIN <- .checkMargin(MARGIN)
+    n <- dim(x)[MARGIN]
+    impVals <- apply(x, MARGIN, quantile, prob = q, na.rm = TRUE)
     for (i in seq_len(n))
         x[is.na(x[, i]), i] = impVals[i]
     x
@@ -467,24 +467,24 @@ impute_MinDet <- function(x, q = 0.01, margin = 2L) {
 ##'
 ##' @param sigma `numeric(1)` controling the standard deviation of the
 ##'     MNAR distribution in `MinProb` and `QRILC`. Default is 1.
-impute_MinProb <- function (x, q = 0.01, sigma = 1, margin = 2L) {
-    margin <- .checkMargin(margin)
+impute_MinProb <- function (x, q = 0.01, sigma = 1, MARGIN = 2L) {
+    MARGIN <- .checkMargin(MARGIN)
     ## transpose if we want to non-default version
-    if (margin == 1L) x <- t(x)
+    if (MARGIN == 1L) x <- t(x)
     x <- imputeLCMD::impute.MinProb(x, q = q, tune.sigma = sigma)
-    if (margin == 1L) x <- t(x)
+    if (MARGIN == 1L) x <- t(x)
     x
 }
 
 ##' @export
 ##'
 ##' @rdname imputation
-impute_QRILC <- function(x, sigma = 1, margin = 2L) {
-   margin <- .checkMargin(margin)
+impute_QRILC <- function(x, sigma = 1, MARGIN = 2L) {
+   MARGIN <- .checkMargin(MARGIN)
     ## transpose if we want to non-default version
-    if (margin == 1L) x <- t(x)
+    if (MARGIN == 1L) x <- t(x)
     x <- imputeLCMD::impute.QRILC(x, tune.sigma = sigma)[[1]]
-    if (margin == 1L) x <- t(x)
+    if (MARGIN == 1L) x <- t(x)
     x
 }
 
@@ -527,23 +527,23 @@ impute_with <- function(x, val) {
 ##' }
 ##'
 ##' impute_matrix(m, FUN = random_imp)
-impute_fun <- function(x, FUN, margin = 1L, ...) {
-    margin <- .checkMargin(margin)
-    if (margin == 2L) x <- t(x)
+impute_fun <- function(x, FUN, MARGIN = 1L, ...) {
+    MARGIN <- .checkMargin(MARGIN)
+    if (MARGIN == 2L) x <- t(x)
     res <- do.call(FUN, list(x, ...))
-    if (margin == 2L) res <- t(res)
+    if (MARGIN == 2L) res <- t(res)
     stopifnot(identical(dim(x), dim(res)))
     res
 }
 
-.checkMargin <- function(margin) {
-    margin <- as.integer(margin)
-    if (!margin %in% c(1L, 2L))
-        stop("'margin' must be 1L or 2L")
-    message("Imputing along margin ", margin,
-            ifelse(margin == 1L, " (features/rows)",
+.checkMargin <- function(MARGIN) {
+    MARGIN <- as.integer(MARGIN)
+    if (!MARGIN %in% c(1L, 2L))
+        stop("'MARGIN' must be 1L or 2L")
+    message("Imputing along margin ", MARGIN,
+            ifelse(MARGIN == 1L, " (features/rows)",
                    " (samples/columns)"), ".")
-    margin
+    MARGIN
 }
 
 ##' @export
@@ -565,7 +565,7 @@ impute_fun <- function(x, FUN, margin = 1L, ...) {
 ##' sapply(ls("package:MsCoreUtils", pattern = "impute_"), getImputeMargin)
 getImputeMargin <- function(fun) {
     args <- formals(fun)
-    i <- grep("margin", names(args))
+    i <- grep("MARGIN", names(args))
     if (length(i)) ans <- args[[i]]
     else ans <- NA
     ans